+++ src/index.ts
@@ -2,8 +2,10 @@ import { Network } from './network'
 import { Contract } from './contract'
 import { Indexer } from './indexer'
 import { Ipfs } from './ipfs'
+import type { Abi as EntryAbi } from './contract/abis/entry/types'
+import type { Abi as PeripheryAbi } from './contract/abis/periphery/types'
 
 export default { Network, Contract, Indexer, Ipfs }
-export { Network, Contract, Indexer, Ipfs }
+export { Network, Contract, Indexer, Ipfs, EntryAbi, PeripheryAbi }
 export * from './types'
 export * as utils from './utils'